Stopped simulation running with more threads than cores with -t argument. Fixed warnings in many files
git-svn-id: https://mac0.cns.ox.ac.uk/svn/SpikeNet/trunk@18 2cedcc20-6066-4341-87df-fb69bdf639a3

+++ Spike/rng.c
@@ -205,7 +205,7 @@ int rands1_new(int mn, int mx, int *iv, int mode)
  Reseed generator if mx <= mn */
 int rnd_new(int mn, int mx)
 {
-	int d, seed;
+	long d, seed;
 	
 	d = mx - mn + 1;
